moved errors to another file
